move helpjson
